Added new api function get comments that returns unlimited comments, added new CommentsComponent. Had to reuse a bunch of code with slight modifications, but that's fine. Nothing too crazy. CommentsComponent just renders the immediate kids and doesn't render sub-comments. For now that's fine. Some day I may implement comment trees.
Also I took out an unnecessary for loop in RenderCardComponent and added
the routes to App.js. The only things left to do now are 1.) theme switcher
and 2.) building a better looking loading... bar, and fixing up loading so
seperate things display loading... on the page instead of everything at once.
Like on user pages Loading... only displays for the user Card and not the users
story posts.

But yeah, more or less it's done and working, so that's great.